requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
python-dotenv
alembic